backend/requirements.txt
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
Pillow
gunicorn
whitenoise
drf-spectacular
drf-spectacular-sidecar